requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Development and Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.0.0

# Utilities
click>=8.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
joblib>=1.3.0
pydantic>=2.0.0
pyyaml>=6.0.0

# File I/O
openpyxl>=3.1.0
xlrd>=2.0.0

# Development Tools
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
